Add unit test to orderSpec and update other files

+++ migrations/sqls/20220109070528-order-videogame-table-up.sql
@@ -1,8 +1,8 @@
 CREATE TABLE order_videogame (
 	id SERIAL PRIMARY KEY,
 	quantity INTEGER,
-	order_id INTEGER REFERENCES orders (id),
-	game_id INTEGER REFERENCES videogames (id)
+	order_id bigint REFERENCES orders (id),
+	game_id bigint REFERENCES videogames (id)
 );
 
 /*store_user_id INTEGER REFERENCES storeusers (id) */
